requirements.txt
Flask
pytesseract
pillow
